Perform repackaging.
