+++ android-app/app/src/main/java/io/spine/examples/todolist/ui/AbstractActivity.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.examples.todolist.lifecycle;
+package io.spine.examples.todolist.ui;
 
 import android.arch.lifecycle.ViewModelProviders;
 import android.os.Bundle;
