+++ android-app/app/src/main/java/io/spine/examples/todolist/ui/newtask/ColorPickerAdapter.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.examples.todolist.task;
+package io.spine.examples.todolist.ui.newtask;
 
 import android.arch.lifecycle.MutableLiveData;
 import android.support.v7.widget.RecyclerView;
@@ -29,7 +29,7 @@ import android.widget.ImageView;
 import com.google.common.collect.ImmutableList;
 import io.spine.examples.todolist.LabelColor;
 import io.spine.examples.todolist.R;
-import io.spine.examples.todolist.Consumer;
+import io.spine.examples.todolist.Callback;
 
 import java.util.List;
 
@@ -88,7 +88,7 @@ final class ColorPickerAdapter extends RecyclerView.Adapter<ColorPickerAdapter.V
      *
      * @param callback the callback triggered when a color is selected
      */
-    void observeColor(Consumer<LabelColor> callback) {
+    void observeColor(Callback<LabelColor> callback) {
         selected.observeForever(callback::accept);
     }
