+++ android-app/app/src/main/res/layout/activity_my_list.xml
@@ -25,7 +25,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="io.spine.examples.todolist.mylist.MyListActivity">
+        tools:context=".ui.mylist.MyListActivity">
     <include layout="@layout/toolbar"/>
     <include layout="@layout/content_my_list"/>
     <android.support.design.widget.FloatingActionButton
