+++ android-app/app/src/main/res/layout/activity_new_task.xml
@@ -26,7 +26,7 @@
         android:layout_height="match_parent"
         android:orientation="vertical"
         android:background="@color/bg_color"
-        tools:context="io.spine.examples.todolist.task.NewTaskActivity">
+        tools:context=".ui.newtask.NewTaskActivity">
     <include layout="@layout/toolbar"/>
     <include layout="@layout/container_new_task"/>
 </LinearLayout>
